refactor(verification): deslop/verbosity cleanup of tutorial files
F1: consolidate five _load_module/_load_tutorial methods into two
    module-level helpers (_import_regressions, _import_tutorial);
    unifies sys.path guard semantics (one variant leaked, one
    double-inserted across a session)

F2: move per-test inline imports (numpy, spectra, BarkerCode,
    generate_bpsk_symbols) to module level; ~15 redundant import
    statements removed, consistent with project test style

F3: remove misleading comment on bpsk_psd_correlation sps=8 that
    claimed the function "does not need to know exactly" while using
    sps directly to derive Rs for the theoretical PSD

F4: add n_fft to results["ofdm"]; main() and _print_summary() now
    reference results["ofdm"]["n_fft"] instead of hard-coding 64

F5: extract n_bits=50_000 variable in test_bpsk_ber_matches_theory
    so the floor expression 1.0/n_bits stays in sync automatically

F6: fix misleading "±1 ± tiny noise" comment in
    test_buggy_bpsk_no_rrc_constellation_is_clean — BuggyBPSK_NoRRC
    adds no noise; samples are exactly ±1+0j

F7: remove three spot-check assertions from test_script_module_importable
    that duplicated exact thresholds already in TestBPSKMeasurements,
    TestOFDMMeasurements, TestBarker13Measurements

F8: fix broaden_pulse kernel dtype from iq.dtype (complex64) to
    float64; moving-average kernel should be real

+++ examples/verification/_tutorial_regressions.py
@@ -80,7 +80,7 @@ def broaden_pulse(iq: np.ndarray, blur_kernel_len: int) -> np.ndarray:
     """
     if blur_kernel_len < 1:
         raise ValueError("blur_kernel_len must be ≥ 1")
-    kernel = np.ones(blur_kernel_len, dtype=iq.dtype) / blur_kernel_len
+    kernel = np.ones(blur_kernel_len, dtype=np.float64) / blur_kernel_len
     return np.convolve(iq, kernel, mode="same").astype(iq.dtype)
